% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mySeqs}
\alias{mySeqs}
\title{Sequences read from fasta file}
\format{
An object of class \code{AAStringSet} of length 10.
}
\source{
\url{https://www.uniprot.org}
}
\usage{
mySeqs
}
\description{
A dataset that contains sequences from ten uniprot IDs, and was read from a fasta file.
}
\details{
@format A AAStringSet of length 10, length of the strings varies:
\describe{
\item{NAMES}{Name of each sequence, in characters}
\item{start}{The start index of each sequence, in integers}
\item{width}{The length of each sequence, in integers}
}
}
\keyword{datasets}
