% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprotToFasta.R
\name{uniprotToFasta}
\alias{uniprotToFasta}
\title{Produces a fasta file that contains the desired Uniprot IDs' sequences.}
\usage{
uniprotToFasta(ID)
}
\arguments{
\item{ID}{A character string that includes the Uniprot IDs separated by "OR"}
}
\value{
Creates a fasta file called seqs.fasta in the main directory.
}
\description{
A function that produces a fasta file based on the Uniprot ID entered
by the users.
}
\examples{
# Use the following UniprotID for this function: P19838, Q00653, Q01201
\dontrun{
uniprotToFasta("AC=P19838 OR AC=Q00653 OR AC=Q01201")
}

# A file called seqs.fasta will appear in the main directory.

}
\references{
\enumerate{
\item Brennan, P. (1970, January 01). Drawing a simple phylogenetic tree of the human rel homology domain family. Retrieved December 04, 2020, from http://rforbiochemists.blogspot.com/2016/12/drawing-simple-phylogenetic-tree-of.html
\item Code for retrieving Uniprot data were borrowed from: http://rforbiochemists.blogspot.com/2016/12/drawing-simple-phylogenetic-tree-of.html
\item Pages, H., Aboyoun, P., Gentleman, R., & DebRoy, S. (2016). Biostrings: String objects representing biological sequences, and matching algorithms. R package version, 2(0), 10-18129.
\item Paradis, E., Claude, J., & Strimmer, K. (2004). APE: analyses of phylogenetics and evolution in R language. Bioinformatics, 20(2), 289-290.
\item Yu, G., Smith, D. K., Zhu, H., Guan, Y., & Lam, T. T. Y. (2017). ggtree: an R package for visualization and annotation of phylogenetic trees with their covariates and other associated data. Methods in Ecology and Evolution, 8(1), 28-36.
\item Charif, D., Lobry, J. R., Necsulea, A., Palmeira, L., Penel, S., Perriere, G., & Penel, M. S. (2020). Package ‘seqinr’.
}
}
