% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprotToPhy.R
\name{uniprotToPhy}
\alias{uniprotToPhy}
\title{Produces the phylogenetic tree graph}
\usage{
uniprotToPhy(ID)
}
\arguments{
\item{ID}{A character string that includes the Uniprot IDs separated by "OR"}
}
\value{
Returns a graph of a phylogenetic tree calculated on the Uniprot IDs
entered along with its full MSA plot.
}
\description{
A function that produces a phylogenetic tree based on the Uniprot ID entered
by the users.
}
\examples{
# Use the following UniprotID for this function: P19838, Q00653, Q01201
uniprotToPhy <- function(ID="AC=P19838 OR AC=Q00653 OR AC=Q01201")
# Access the phylogenetic tree with MSA plot directly.
uniprotToPhy

}
