% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqToPhy.R
\name{seqToPhy}
\alias{seqToPhy}
\title{Produces the phylogenetic tree graph from sequence file}
\usage{
seqToPhy(mySeqs)
}
\arguments{
\item{mySeqs}{A AAStringSet of sequences, length of each sequence varies.}
}
\value{
Returns a graph of a phylogenetic tree calculated on the sequence dataset
along with its full MSA plot.
}
\description{
A function that produces a phylogenetic tree based on the fasta file uploaded
by the users.
}
\examples{
# Use the mySeqs dataset for this function:
uniprotToPhy <- function(mySeqs)
# Access the phylogenetic tree with MSA plot directly.
uniprotToPhy

}
