% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqToPhy.R
\name{seqToPhy}
\alias{seqToPhy}
\title{Produces the phylogenetic tree graph from sequence file}
\usage{
seqToPhy(mySeqs)
}
\arguments{
\item{mySeqs}{A AAStringSet of sequences, length of each sequence varies.}
}
\value{
Returns a graph of a phylogenetic tree calculated on the sequence dataset
along with its full MSA plot.
}
\description{
A function that produces a phylogenetic tree based on the fasta file uploaded
by the users.
}
\examples{
# Use the mySeqs dataset for this function:
seqToPhy <- function(mySeqs)
# Access the phylogenetic tree with MSA plot directly.
seqToPhy

}
\references{
Code for retrieving Uniprot data were borrowed from: http://rforbiochemists.blogspot.com/2016/12/drawing-simple-phylogenetic-tree-of.html
Pages, H., Aboyoun, P., Gentleman, R., & DebRoy, S. (2016). Biostrings: String objects representing biological sequences, and matching algorithms. R package version, 2(0), 10-18129.
Paradis, E., Claude, J., & Strimmer, K. (2004). APE: analyses of phylogenetics and evolution in R language. Bioinformatics, 20(2), 289-290.
Yu, G., Smith, D. K., Zhu, H., Guan, Y., & Lam, T. T. Y. (2017). ggtree: an R package for visualization and annotation of phylogenetic trees with their covariates and other associated data. Methods in Ecology and Evolution, 8(1), 28-36.
}
